+++ mobile/android/android-components/components/browser/tabstray/src/test/java/mozilla/components/browser/tabstray/DefaultTabViewHolderTest.kt
@@ -4,27 +4,32 @@
 
 package mozilla.components.browser.tabstray
 
+import android.content.res.ColorStateList
 import android.graphics.Bitmap
+import android.graphics.drawable.ColorDrawable
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import mozilla.components.concept.base.images.ImageLoadRequest
+import mozilla.components.concept.base.images.ImageLoader
 import mozilla.components.concept.tabstray.Tab
 import mozilla.components.concept.tabstray.TabsTray
 import mozilla.components.support.base.observer.ObserverRegistry
-import mozilla.components.concept.base.images.ImageLoadRequest
-import mozilla.components.concept.base.images.ImageLoader
 import mozilla.components.support.test.any
 import mozilla.components.support.test.eq
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.nullable
 import mozilla.components.support.test.robolectric.testContext
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNull
+import org.junit.Assert.assertSame
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.never
+import org.mockito.Mockito.spy
 import org.mockito.Mockito.verify
 
 @RunWith(AndroidJUnit4::class)
@@ -181,5 +186,78 @@ class DefaultTabViewHolderTest {
         viewHolder.bind(tab, false, mock(), mock())
 
         assertEquals(drawable, thumbnailView.drawable)
+    }
+
+    @Test
+    fun `bind sets the values for this instance's Tab and TabsTrayStyling properties`() {
+        val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
+        val viewHolder = DefaultTabViewHolder(view)
+        val tab = Tab("123", "https://example.com")
+        val styling: TabsTrayStyling = mock()
+
+        assertNull(viewHolder.tab)
+        assertNull(viewHolder.styling)
+
+        viewHolder.bind(tab, false, styling, mock())
+
+        assertSame(tab, viewHolder.tab)
+        assertSame(styling, viewHolder.styling)
+    }
+
+    @Test
+    fun `bind shows an item as selected or not`() {
+        val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
+        val tab = Tab("123", "https://example.com")
+        val viewHolder = spy(DefaultTabViewHolder(view))
+
+        viewHolder.bind(tab, false, mock(), mock())
+        verify(viewHolder).updateSelectedTabIndicator(false)
+
+        viewHolder.bind(tab, true, mock(), mock())
+        verify(viewHolder).updateSelectedTabIndicator(true)
+    }
+
+    @Test
+    fun `updateSelectedTabIndicator should further delegate to the appropriate method`() {
+        val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
+        val viewHolder = spy(DefaultTabViewHolder(view))
+
+        viewHolder.updateSelectedTabIndicator(showAsSelected = true)
+        verify(viewHolder).showItemAsSelected()
+
+        viewHolder.updateSelectedTabIndicator(showAsSelected = false)
+        verify(viewHolder).showItemAsNotSelected()
+    }
+
+    @Test
+    fun `showItemAsSelected should use TabsTrayStyling for indicating that an item is currently selected`() {
+        val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
+        val viewHolder = spy(DefaultTabViewHolder(view))
+        val tab = Tab("123", "https://example.com")
+        val styling = TabsTrayStyling()
+
+        // Need to be called first to set the styling for this holder
+        viewHolder.bind(tab, false, TabsTrayStyling(), mock())
+        viewHolder.updateSelectedTabIndicator(true)
+
+        assertEquals(styling.selectedItemTextColor, viewHolder.titleView.textColors.defaultColor)
+        assertEquals(styling.selectedItemBackgroundColor, (viewHolder.itemView.background as ColorDrawable).color)
+        assertEquals(ColorStateList.valueOf(styling.selectedItemTextColor), viewHolder.closeView.imageTintList)
+    }
+
+    @Test
+    fun `showItemAsSelected should use TabsTrayStyling for indicating that an item is not currently selected`() {
+        val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
+        val viewHolder = spy(DefaultTabViewHolder(view))
+        val tab = Tab("123", "https://example.com")
+        val styling = TabsTrayStyling()
+
+        // Need to be called first to set the styling for this holder
+        viewHolder.bind(tab, true, TabsTrayStyling(), mock())
+        viewHolder.updateSelectedTabIndicator(false)
+
+        assertEquals(styling.itemTextColor, viewHolder.titleView.textColors.defaultColor)
+        assertEquals(styling.itemBackgroundColor, (viewHolder.itemView.background as ColorDrawable).color)
+        assertEquals(ColorStateList.valueOf(styling.itemTextColor), viewHolder.closeView.imageTintList)
     }
 }