+++ mobile/android/android-components/components/browser/tabstray/src/test/java/mozilla/components/browser/tabstray/TabsAdapterTest.kt
@@ -7,6 +7,8 @@ package mozilla.components.browser.tabstray
 import android.view.View
 import android.widget.FrameLayout
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import mozilla.components.browser.tabstray.TabsAdapter.Companion.PAYLOAD_DONT_HIGHLIGHT_SELECTED_ITEM
+import mozilla.components.browser.tabstray.TabsAdapter.Companion.PAYLOAD_HIGHLIGHT_SELECTED_ITEM
 import mozilla.components.concept.tabstray.Tab
 import mozilla.components.concept.tabstray.Tabs
 import mozilla.components.concept.tabstray.TabsTray
@@ -17,6 +19,8 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.ArgumentMatchers
+import org.mockito.Mockito.never
 import org.mockito.Mockito.spy
 import org.mockito.Mockito.verify
 
@@ -28,6 +32,8 @@ private class TestTabViewHolder(view: View) : TabViewHolder(view) {
         styling: TabsTrayStyling,
         observable: Observable<TabsTray.Observer>
     ) { /* noop */ }
+
+    override fun updateSelectedTabIndicator(showAsSelected: Boolean) { /* noop */ }
 }
 
 @RunWith(AndroidJUnit4::class)
@@ -134,5 +140,63 @@ class TabsAdapterTest {
         )
 
         verify(observer).onTabsUpdated()
+    }
+
+    @Test
+    fun `onBindViewHolder will use payloads to indicate if this item is selected`() {
+        val adapter = TabsAdapter()
+        val holder = spy(TestTabViewHolder(View(testContext)))
+        adapter.updateTabs(Tabs(listOf(mock(), mock()), selectedIndex = 1))
+
+        adapter.onBindViewHolder(holder, 0, listOf(PAYLOAD_HIGHLIGHT_SELECTED_ITEM))
+        verify(holder, never()).updateSelectedTabIndicator(ArgumentMatchers.anyBoolean())
+
+        adapter.onBindViewHolder(holder, 1, listOf(PAYLOAD_HIGHLIGHT_SELECTED_ITEM))
+        verify(holder).updateSelectedTabIndicator(true)
+    }
+
+    @Test
+    fun `onBindViewHolder will use payloads to indicate if this item is not selected`() {
+        val adapter = TabsAdapter()
+        val holder = spy(TestTabViewHolder(View(testContext)))
+        adapter.updateTabs(Tabs(listOf(mock(), mock()), selectedIndex = 1))
+
+        adapter.onBindViewHolder(holder, 0, listOf(PAYLOAD_DONT_HIGHLIGHT_SELECTED_ITEM))
+        verify(holder, never()).updateSelectedTabIndicator(ArgumentMatchers.anyBoolean())
+
+        adapter.onBindViewHolder(holder, 1, listOf(PAYLOAD_DONT_HIGHLIGHT_SELECTED_ITEM))
+        verify(holder).updateSelectedTabIndicator(false)
+    }
+
+    @Test
+    fun `onBindViewHolder with payloads will return early if there are currently no open tabs`() {
+        val adapter = TabsAdapter()
+        val holder = TestTabViewHolder(View(testContext))
+        val payloads = spy(arrayListOf(PAYLOAD_DONT_HIGHLIGHT_SELECTED_ITEM))
+
+        adapter.onBindViewHolder(holder, 0, payloads)
+        // verify that calls we expect further down are not happening after the null check
+        verify(payloads, never()).isEmpty()
+        verify(payloads, never()).contains(ArgumentMatchers.anyInt())
+
+        adapter.updateTabs(Tabs(emptyList(), selectedIndex = 0))
+        adapter.onBindViewHolder(holder, 0, payloads)
+        // verify that calls we expect further down are not happening after the null check
+        verify(payloads, never()).isEmpty()
+        verify(payloads, never()).contains(ArgumentMatchers.anyInt())
+    }
+
+    @Test
+    fun `onBindViewHolder with empty payloads will call onBindViewHolder and return early for a full bind`() {
+        val adapter = TabsAdapter()
+        val holder = TestTabViewHolder(View(testContext))
+        val emptyPayloads = spy(arrayListOf<String>())
+        adapter.updateTabs(Tabs(listOf(mock()), selectedIndex = 0))
+
+        adapter.onBindViewHolder(holder, 0, emptyPayloads)
+
+        verify(emptyPayloads).isEmpty()
+        // verify that calls we expect further down are not happening after the isEmpty check
+        verify(emptyPayloads, never()).contains(ArgumentMatchers.anyString())
     }
 }